rename skyobjects (plural) to skyobject (singular)
